Move osm request to osm
